Fix bug on stuck for Keepalived signal response
In some times exporter will stuck because of high response time of signals and prometheus will stuck too.
Now we have set a timeout of 1sec for getting response from Keepalived and if there were no response from Keepalived we will proceed with previous data if it has dumped and we will wait for Keepalived to dump stats in another goroutine.